Move `get_all_modules()` to tools.py
